Fix wrong academic year being shown
When displaying patients in a different academic year to the current
one, that academic year should be used to determine the patient year
group. Instead, in a few places the wrong academic year was being used.

I've also added a feature test that covers the scenario of uploading
patients during the preparation period for the upcoming academic year to
make sure this feature is working as expected and doesn't regress.