Replace DraftClassImport with DraftImport
This replaces the existing draft class import model with a model that
handles more of the steps of the import process, specifically it now
includes the step of choosing the type of import.

I've made this change because we're probably going to adding a step to
two of the import types allowing users to specify which academic year
they're importing records for, and therefore we need this page to be
shown for multiple import types.

By making this change it also means we can get validation if an invalid
type is specified.

Jira-Issue: MAV-1700