Add academic_year column to imports
This adds a new `academic_year` column to imports which allows us to
create imports during the preparation period that target a specific
academic year, rather than currently where it's always the pending
academic year.

We don't anticipate that this will be used much by SAIS teams, but
it's useful for testing.

Jira-Issue: MAV-1700