Use patient_id when staging vaccination record changes
This updates the immunisation import process to ensure that when
importing details for a vaccination record that already exists, we use
the `patient_id` as the value rather than the `patient` object. This
ensures that the resulting pending changes can be saved successfully as
it contains a single value rather than a hash of the patient.

Although we anticipate this structure will change soon with the PDS
improvements, we saw an unhandled error occur in the test environment so
this puts in a fix for the issue.

Sentry-Issue: 6784604544