feat: add expiration event for seasonal programmes' health information
Even when a child was vaccinated for these programmes, the health
information still expires. New consent must be sought for the next
academic year. In fact, seasonal programmes are the reason we introduced
expiry of health information in the first place.

A custom iterator on the AcademicYear class is made to avoid the linter
failing on .all.each as this chaining is an anti-pattern when used on
ActiveRecords. https://rails.rubystyle.guide/#find-each