Rename unprocessable_entity
In Rack 3.2, this has been renamed to unprocessable_content and using
the old symbol is deprecated, which results in a warning message.

Although a future version of Rails will handle this for us
(https://github.com/rails/rails/pull/53383) but we can also change the
symbol we use in our codebase to reduce the number of warnings happening
at the moment.