Redirect to patient if no session exists
If a patient is being matched to a consent form manually, and the
patient is not eligible for any of the programmes we can't redirect the
user to the patient in a session as it fails to generate a suitable URL.
Instead we can redirect the user to the global child record.

This is unlikely to happen in production, but we've been seeing it
happen a few times in the test environments due to test data.

Sentry-Issue: 6591726028