Don't add school patients to community clinic
This updates the logic around school moves to not add patients in to the
community clinic if they're assigned to a school. This is to reduce the
number of patients who will exist in the community clinic session, to
make it more useable, especially when it comes to flu where the total
number of patients is high.

Jira-Issue: MAV-1430