Add CIS2Info
This adds a new model class which is backed by the browser session that
stores all the information related to CIS2 authentication. Currently
this information is stored in a hash, but by having a model class we can
access the data in a more structured way allowing for code that's easier
to comprehend.

Jira-Issue: MAV-1280