Show year group according to pending academic year
In https://github.com/nhsuk/manage-vaccinations-in-schools/pull/4175 we
made it so that the year group is hidden for some patients. This was
supposed to apply for the pending academic year, not the current one.
The idea is that during the preparation period, you would be looking at
patients who are due to age out at the start of next academic year, and
for those patients the year group is hidden.

Jira-Issue: MAV-1512