Ensure patients are unarchived
When an archived patient is re-imported, we should unarchive them. This
was added in to the school moves in
b35cc281ea7e813f5525ee4d52f1a17049fab6ba however, we didn't handle the
case where a school move wouldn't be created.

If a patient is archived, and they're being re-imported exactly as they
were (same school, etc) we don't create a school move. This results in
them not being unarchived.

Instead, we can ensure that a school move is created if the patient is
already archived.

I haven't written a feature test for this scenario as it's being covered
by the regression tests.

Jira-Issue: MAV-1506