Send clinic invitations to all patients
In the previous commits we can restored the old behaviour where patients
are only added to the community clinic manually. This means we can
partially revert 46456dc62bb7c2064eac3ec9448fe6fd90fa61ed and
automatically send clinic invitations to all patients in the clinic.

This also allows means there's a bit of time between inviting patients
to the clinic, and the email going out, allowing for patients to be
recorded as vaccinated and avoid an unnecessary invitation email.

This is less relevant for inviting lots of patients to the clinic after
a session is completed, but it does become relevant when adding
individual patients to a session if they turn up at a session.

Jira-Issue: MAV-1430
Jira-Issue: MAV-278