Hide patients who don't need consent
This updates the "Consent" tab on a session page to hide patients who
don't need consent. This can occur because the patient is already
vaccinated for the programme this year, and therefore should no longer
be shown on this tab.

Instead they can continue to be found in the "Session outcomes" tab
(which is shortly to be renamed to "Children").

Jira-Issue: MAV-1679