Rename "Session outcomes" tab
This renames the "Session outcomes" tab to "Children" and moves it
immediately to the right of the "Overview" tab. The intention behind
this change is to make this the tab that users use to find children in a
session as we will be hiding patients from the consent tab in the
future.

Jira-Issue: MAV-1679