Add status generators
This adds a new set of classes (and refactors how the `Status`-prefixed
models work) to use these new classes. The reason to do this is so that
the logic to determine the status of a patient can be used outside of
the model classes themselves. More specifically, we can use them as part
of generating statuses, as a few of the statuses are interdependent.

For example, at the moment to determine the triage status you need to
konw the vaccination status and the consent status.

Similarly, this commit will be followed up by a commit that uses the
session status to determine the overall programme status.

Jira-Issue: MAV-1679