Display latest session outcome
When viewing a programme outcome, we want to show the latest session
outcome next to it, so users understand the reason behind the patient
having a particular status.

For example, if a patient has not yet been vaccinated, because they were
unwell during the session, we show that label next to the overall
outcome.

Jira-Issue: MAV-1679