Fix strict loading error when viewing patient card
This issue was introduced by adding the `no_notify` important notice, and the accompanying warning added to a patient card (added in [#4176](https://github.com/nhsuk/manage-vaccinations-in-schools/pull/4176)).

Because this notice (and warning) is dynamic, and needs to access `VaccinationRecord.programme`, it was causing a strict loading error. Adding some extra `includes` blocks in the relevant controllers has resolved the issue.