Remove vaccine_methods assignment
This was added in e63f3d0d4164e2f9616a8d2a1241b27de7127952 when we first
started recording vaccine method on consent forms, and it was defaulted
to `injection` as we didn't have support for nasal flu at the time.

Since then, the users are now asked which vaccine method they want to
consent to, and we don't need to specify the vaccine method upfront.
Either way, this was doing no harm as it was being reset by the
`reset_unused_attributes` method anyway. However, I think it looks
strange and could be potentially confusing.