Add ReportingApi::OneTimeToken model (+ spec, factory)
This is an 'under the hood' change which just adds the model.
It's not actually used anywhere yet. A future PR will use it when
authenticating users of the reporting app, but that app is yet to
be written

(this PR was split out of the larger spike in MAV-1406, PR #3866)