Fix specs for consent and invitation jobs
These tests weren't working because they set up a session in 3 weeks
time, and that meant when running the tests after today (the 11th
August), the sessions would be set up for a different academic year to
the consent requests and other associated models created today.

This fixes the tests by specifying a time and running the tests within a
frozen time.