Skip selecting a team if there is only one
This improves the user journey of signing in by avoiding the need to
choose a team, if there is only one team available to choose from. This
ensures that for most users they won't see any difference when this PR
is deployed. Only the users who are required to choose a team will see
the form.

Jira-Issue: MAV-1280