Don't check patient pending changes on vaccination record
The current scope on vaccination records means that if a patient has
import issues, we end up seeing a row saying the vaccination record
itself has an import issue, which is not correct, it should instead be
just the patient.

This scope was introduced in 12ea88fa927133cd42778730bda9a2617f2713ff
and the logic around imports has changed significantly since then.
Specifically, we now doesn't apply pending changes to patients when
importing vaccination records which was the reason for this scope in the
first place.

Instead, the import issues controller combines issues across vaccination
records and patients at a higher level.

The same tests pass before and after demonstrating that we're not losing
any functionality from this change.