Only show aged out filter in global child view
This shouldn't be shown in sessions because you won't ever see patients
who have aged out of the programme in sessions. The original intention
behind the design was to show it only on the children page.

Currently if you try and use this filter you get an error message.

Sentry-Issue: 6803159132