fix: hide 'register attendance' when session not accepting registration
It is not possible to register a child when the session is not in progress
as such, we should not show the row + call to action when the session does
not happening today.

The `if session.requires_registration` was moved so that all visibility
checks are performed in the same place.