Use pending academic year instead of current
When viewing a patient outside the context of a session, this renders
the information for a patient in the context of the pending academic
year rather than the current one. This is to make the preparation period
easier for users by showing year groups and sessions as they will be
starting in September.