Hide patients vaccinated by the same organisation but different team
At the moment if a patient is vaccinated by one team, and there's
another team in the same organisation, they will be shown the patient
because it's vaccination record has the same ODS code.

Instead, for a non-historical vaccination, we shouldn't be looking at
the ODS code, instead we should be looking at the team that's attached
to the session.

Jira-Issue: MAV-1280