Allow signing in without a team workgroup
This allows users to sign in to any team that's part of an organisation,
even if they're not part of the workgroup for that team.

This is necessary to allow release 2.17 to be deployed even if the SAIS
teams haven't yet updated their CIS2 workgroups. This change will need
to be reverted in 2.18.