Handle unknown session keys
This ensures that when creating an instance of a
`RequestSessionPersistable` model, if there are unknown keys in the
session data these get ignored rather than the current behaviour where
an exception is raised.

This allows us to migrate attributes in these models without needing
users to sign out and clear their browser session data, instead any old
values will simply be ignored.