Add Child import PDS extravaganza test
This will be used to test the different scenarios that relate to PDS
lookups being done as part of the import process. For starters, though,
it just confirms that when no fields are changed, the is no pending
changes saved to a patient. This test appears to have been missed
elsewhere.