Add script to prepare East of England for flu
This prepares the East of England organisation for flu and rollover
starting next week. Specifically this set the CIS2 workgroup for their
team, adds the missing flu schools and adds the flu programme.

This doesn't create any flu sessions as that will happen on Monday
through the use of the `create-sessions` command line tool.

Jira-Issue: MAV-1120
Jira-Issue: MAV-1664