Remove invalid nhs number test from extravaganza
This was testing a very narrow edge-case which we will probably
shut-down in a following PR.

Specifically, when we find a patient using
their NHS number with PDS, and it is invalid, this was testing that the
import would invalidate the patient. However this is unlikely to happen
since the patient would have already been invalidated as a result of the
daily PDS update job. Additionally, in the future we'll remove the
find-by-nhs-number lookup as part of the new PDS lookup flows.

Jira-Issue: MAV-1390