Update extravaganza test for PDS search only
This supports the new ProcessChangesetsJob that only looks up NHS numbers
for patients not uploaded with NHS numbers, and gets rid of PDS find.
These tests check for creating a new patient not uploaded with an NHS number, but has found one through PDS.
And not doing any PDS lookups when uploaded with an NHS number.