Increase PDS rate limit to 50 req/s
Our PDS rate limit will be increased to 50 req/s on Monday 18th August
by the end of the day so need to make the change in our code to make
requests at a higher rate.

This is set as a setting value so we can dynamically change this value.