Migration to add PDSSearchResults table
This will store the results from PDS searches done during a class or cohort import.
It will store the step name, NHS number retrieved if any and result (one match, too many etc).

We can use this to surface which step found the NHS number on a record to the user in the future.