Use SystmOne code in exporters and importers
When exporting or importing data to or from a SystmOne file, we need to
use the SystmOne code if possible. This is to support cases where the
location code in SystmOne doesn't match the URN or ODS code in Mavis.

Specifically, we know this occurs in the case of school sites with
multiple campuses each sharing the same URN but with very different
names. In Mavis we have a site code for each of these locations, but the
value in SystmOne will be different.

Jira-Issue: MAV-1384