Rename `is_nurse?`
To `can_prescribe_pgd?`. This renames the method to move towards a
world where we check what permissions a user has to take certain
actions rather than basing their access permissions on their role.

This is needed because in the future we will have a prescriber role,
which can have any CIS2 role and will instead be identified by their
activities.

Jira-Issue: MAV-1680