Don't check for nurse or admin
We already do this check when we authenticate the user before each
request, so we don't need to check this each time an action is
performed.

This is necessary because we will have more user types in the future,
and in particular we'll have users with any kind of role, so we can't
check every possible role. Instead, the authentication will check that
the users have permission to form an action.