Fix flaky consent matching test
We were previously searching for patients based on their first name,
which in most cases will return a response, but if the randomly
generated first name is a lot shorter than the last name then the search
won't match with the patient.

Instead we can search based on the full name to guarantee a result.