Handle `nil` vaccination record location
In the SystmOne exporter. This situation can occur if a vaccination is
recorded in the community clinic, where the `location` will be `nil`,
and the `location_name` will be set instead.

This change was made in aae7b838a035e4d7124b95c80702f29a1a1fd1f9,
previously vaccination records in the generic clinic would have a
location of the generic clinic, but we want to move to a world where the
location is the specific community clinic where the record took place.

This error was spotted by the regression tests and fixes this Sentry
error: https://good-machine.sentry.io/issues/6818602795/