Add function to make `search` requests
`NHS::ImmunisationsAPI::search_immunisations` sends the search request to the API, and handles the response.

NB: There is a bug in the API code which leads to the `Bundle.link` being incorrect. I have implemented a solution which will accept both the current incorrect version, as well as the fixed version after they deploy their fix. After that, the extra logic in `check_bundle_link_params` can be removed.