Flatten nested `EXISTS`
This refactors the way the `PatientSession#appear_in_programmes` scope
works to flatten the query which currently has an `EXISTS` inside an
`EXISTS` as the logic still works when the nesting is flattened, and we
think this should have a performance boost to this query.

I haven't been able to demonstrate the performance increase as we don't
have good analysis available on our production database, but it's
unlikely to make it any slower so I think we can safely apply this
change.