Do not allow "keep both records" when match is on NHS number
Currently, we store any import differences that need to be flagged for review by a user to `pending_changes`. These accumulate over susequent imports. Until we can make changes specific to an import, I have implemented it so that the last changeset on a patient is used to determine whether or not a patient matched an existing patient on NHS number at the time of upload. Of course, this isn't strictly correct as an old uplaod that caused pending changes ont he patient, might not have matched on NHS number if it was done a) before MAV-1788 was live and b) the duplicate was not reviewed.

However, we expect teams to clear out their import issues after every upload and we will ensure there are no pending changes before deploying this. So whether or not a patient matched on NHS number will be true.