Ensure only one school move exists per patient
This updates the indexes on the school move table to ensure that only
one school move can exist per patient, and makes it so that when you
import a new school move it replaces any that already exist. This is to
resolve the current confusing behaviour where accepting a school move
deletes any others, so users need to know which order to accept school
moves in.

Jira-Issue: MAV-1806