Add delegation columns to sessions
As part of the delegation work, sessions will need a boolean field to
determine if the session is using the PSD protocol or national
protocol. PSD or Patient Specific Direction is the protocol that allows
healthcare workers to administer nasal vaccines without having
the necessary qualifications as a nurse.

Setting the default value to false until the full delegation feature
is ready to go live.