Set `location_id` on vaccination records in clinic
This ensures that when we're recording a vaccination from a community
clinic session, we link the location of the community clinic or school
using the `location_id` foreign key rather than storing the name of the
location in `location_name`. This enables us to have more structured data,
including access to various other bits of information on the location.

We still need the `location_name` column to handle importing unknown
locations.

Jira-Issue: MAV-550