Ensure location URNs are unique
In 8a0d5e1ed2be5c9f407fad45297c7ca84e542fd2 a new `site` column was
added and a unique index was added against the `site` and `urn`
combination. However, for most locations there won't be a site, instead
it will be left `NULL`. In that case, the unique constraint fails as
`NULL` values are considered distinct, meaning it passes the unique
constraint.

Instead we need to add back the index to ensure that when on their own,
URNs are unique.

Jira-Issue: MAV-1384