Replace Gillick assessment patient session foreign key
This replaces the foreign key association between Gillick assessments
and patient sessions to instead link directly to the patient and the
session (via the date). This is needed as we eventually want to replace
the `PatientSession` model and to do that we need to make sure all
foreign keys to it have been replaced.

The functionality should be the same before and after, actually slightly
improved as Gillick assessment is only supposed to exist on the day it
happened, so now we link directly to the session date rather than the
overall session.

Jira-Issue: MAV-1818