Validate child names on import
PDS restricts the allowable characters in a patient's name to be
alphanumeric, including letters used in European alphabets, and the
extra characters '-', '.' and "'". Especially now that we perform PDS
updates as a part of the import process, we need to ensure that we
follow these same guidelines that PDS uses.

Jira-Issue: MAV-1807