Replace session attendance patient session foreign key
This replaces the foreign key association between session attendances
and patient sessions to instead link directly to the patient. This is
needed as we eventually want to replace the PatientSession model and to
do that we need to make sure all foreign keys to it have been replaced.

The functionality should be the same before and after.

Jira-Issue: MAV-1820