Add `sessions configure` command
This adds a new command to the CLI which allows ops to bulk enable or
disable delegation options or registration requirement for all the
sessions for a particular team. This is necessary to avoid needing the
teams to manually edit these options for each session (when they may
have hundreds).

Jira-Issue: MAV-1668
Jira-Issue: MAV-1787