Track which users should be shown in list of suppliers
Our users log in using CIS2, and it's only at the point of logging in do
we know what permissions a user has, so at that point we will update the
this value for them. For non-CIS2 environments, we could in theory rely
on the `fallback_role` column instead, but I think to make the code
consistent it makes sense to use same column in both cases.

Jira-Issue: MAV-1353