Fix labelling of Gillick assessments in activity log
The activity log labels each assessment as either "Completed" or
"Updated" depending on if it's the first one, and this means we need to
guarantee the order as it works by checking if we're processing the
first entry.

This also fixes a flaky feature test that depends on the right label
appearing correctly.