fix: prevent uploads of excessively old patients
Anyone born before 2000 is clearly not going to be in secondary education 
anymore. This isn't a perfect cut-off point but it's low enough that it wil
never be a problem for anyone who still is going to school, while it will 
prevent uploads of children born in the wrong century, which we sometimes see
when the CHILD_YEAR_GROUP parameter is set while the date of birth for example 
does not have a century set e.g. 01/01/0015 may show as 01/01/15 in excel

Jira-Issue: MAV-1833