Add `Session#pgd_supply_enabled?`
This adds a method which can be used to determine whether PGD supply is
enabled for a session. At the moment this works the same as whether
delegation is supported, as we know that every SAIS team supports PGD
supply delegation, however by having this method it allows us to change
this in the future easily.