Prevent deleting sessions if Gillick assessment was performed
This updates the criteria for when a session date can be deleted to
ensure that if a Gillick assessment exists for that date, it should not
be deleted as that means that a patient was seen on that date.

Previously it was possible to delete session dates as they didn't have a
foreign key to the Gillick assessment, but with the recent refactor they
now do and therefore the UI needs to be updated to match.

Jira-Issue: MAV-1818