Fix `session_attendance` association
This fixes an issue where the `session_attendance` association on the
`PatientSession::RegistrationStatus` ends up being attached to the wrong
record when pre-loaded. As far as I can tell this appears to be a bug in
Rails as I can't see anything wrong with the approach, but the test I've
written reproduces the problem. When using `eager_load` the correct
association is found, but when using `preload` the criteria for the
session attendance to be today's atendance seems to be lost. Having
investigated the queries a little, it seems to be due to the fact that
the association is through `patient_session`.

The impact of this issue means that if a patient has multiple session
attendance records (i.e. they have attended or not attended multiple
session dates), then sometimes the status won't get updated correctly.