Always run imports in the background
With the new background process where PDS checks are performed, we're
seeing a problem where if any of the PDS checks fail due to a 429 error,
the users see an unexpected error page because we catch the error
expecting Sidekiq to re-process it. However, if the user uploads a file
that is smaller than 15 rows we currently process it synchronously
meaning we can't take advantage of the retry behaviour in the job.

Rather than trying to make the synchronous process work well with
retries, it seems simpler to make all imports run in the background,
especially since in practice the vast majority of files are larger than
15 rows.