Refactor waiting for imports to finish
This improves how the feature tests wait for imports to finish by
introducing a method, `wait_for_import_to_complete`, which can be used
after uploading a file to ensure that all the related jobs have
processed before moving on to the other parts of the test.

I've also removed the slow cohort import as that was designed to test
the background processing, but now all imports are processed in the
background so this saves a bit of time running the tests.