Add within_academic_year rspec helper
During the period between academic years, the preparatory period, some
validations like on session dates may fail. This helper allows us to
easily configure a spec to run within the academic year by specifying
`:within_academic_year` on a test:

  it "does something", :within_academeic_year do

this will travel to the beginning of the academic year, 1st of Sept. It
can also be configured to start later than the first day with the
"from_start" parameter:

  it "does something", within_academeic_year: { from_start: 7.days } do