Allow HCAs to choose injection under national protocol instead of PSD
In the scenario where national protocol and PSD is enabled for a flu
session and the patient has consented for nasal or injection, the HCA
can choose injection as a fallback option.

In any case, it's necessary for the HCA to provide the supplier of the
vaccine if they want to go with injection. The vaccination record will
record it as done through national protocol.