Trigger an Imms API search on NHS number change
When a patient's NHS number is changed, the records which had previously been pulled from the Imms API will no longer be valid as they're associated with a patient with a different NHS number.

This way, when the NHS number is changed, the old records will get deleted, and replaced with whichever records are returned in the latest search response for the new NHS number.