Add `read_immunisation_by_nhs_immunisations_api_id`
This makes the API call to read a vaccination record in the API by passing in an API ID string. It returns a `FHIR::Immunization` object.

This is useful because it will allow operational tasks to discover records in the API which don't have a corresponding `VaccinationRecord` object in Mavis