Make `read_immunisation` a thin wrapper around `read_immunisation_by_nhs_immunisations_api_id`
This reduces code duplication a lot, while maintaining the same functionality.

The only downside with this is that the error and log messages will be slightly different ("reading by ID" vs "reading record #{ID}"). Given that we only expect these functions to be run by operational staff, this doesn't seem like too much of a problem.