Ensure PGD supply is disabled when PSD is enabled
This refactors various parts of the code related to delegation to ensure
that when PSD is enabled, PGD supply is disabled. Specifically, this
fixes a number of code paths which weren't checking if
`pgd_supply_enabled?` was `true` (and instead assumed that it was always
true) and also updates the filtering on the record tab to consider the
case when PSD and national protocol is enabled.

I had to get rid of the `vaccine_method_for` method as it doesn't work
in the case of PSDs, where the vaccine method can be different per
patient.

I've also fixed a bug where HCAs are able to record injection vaccines
if PGD supply is enabled and national protocol is not.

Jira-Issue: MAV-1893
Jira-Issue: MAV-1895