Ensure PSDs are created after verbal consent
When a prescriber records verbal consent, we also ask the user if they
would like to add a PSD to the patient. This wasn't working correctly,
and instead a PSD wasn't being added to the patient. This is because the
logic related to creating PSDs was specific to the triage journey that
occurs outside of the verbal consent journey. Instead I've moved the
logic in to the `TriageForm` ensuring that both triage journeys will
have the same logic.

Jira-Issue: MAV-1905