Avoid job failures for team API key
In test and QA we send emails using team API keys, which fails for most
users but will allow emails and SMSs to be sent to users in the guest
list. This change will prevent these failures from causing the jobs to
be retried and instead the job is considered a success.