Create database migration to enable feature flags
When this release is deployed, the new feature flags will default to
being disabled. This has the potential to create issues;
if there are any vaccination records created before they are manually
enabled, then they won't be synced. This way, the functionality will 
be enabled immediately after the new version goes live, meaning this is
no longer a concern.