Ensure programme tags are ordered
This ensures that when displaying a list of programme tags they are
always ordered alphabetically, even if they're not ordered in that way
when they're passed in.

Generally we order the results of programmes when querying from the
database, but there seems to be an issue in QA currently where
preloading the programmes means they are not always ordered. This change
guarantees that users will see the programmes in the correct order.