Decrease GOV.UK Notify rate limit
In October we're going to be rate limited to 50 requests per second, so
to pre-empt this we can change the rate limit on our side early to
ensure that this will work well when the change is made on their side.

I've chosen 45 instead of 50 to ensure we're well under the limit and
won't risk any inadvertant rate limiting errors.

Jira-Issue: MAV-1937