Fix HCAs unable to record vaccination if PSD exists
If a PSD exists on a patient from when the session had PSD enabled, and
then PSD is subsequently disabled, HCAs are unable to record
vaccinations for these patients. This is because the service thinks
we're using the PSD to record the vaccination and doesn't display the
field asking the HCA to specify who supplied the vaccine. Instead, even
if the patient has a PSD, when PSD is disabled on the session, the HCA
will not be using the PSD protocol, they will be using PGD with supply.