Display tallies on record vaccinations tab
This commit introduces a new feature that displays vaccination tallies on
the "Record vaccinations" tab of a session. This provides nurses with a quick
overview of the number of vaccinations they have administered during the session.

The tallies are specific to the logged-in and addresses the need for a more
efficient way to track administered vaccinations

Additionally, this commit merges the UI for choosing the "default batch" with the
new vaccination tallies UI, streamlining the user experience.