Trigger an Imms API search during bulk imports
When `Patient.import` is called, the `after_update` hooks are not called.
As such, we need to explicitly check whether the NHS number has changed,
and if it has, enqueue the search job.

This functionality has not been feature tested because we believe that it's
very much an edge-case. If it becomes an issue, we can add the relevant
testing.