Ensure emails don't end with a dot
If an email ends with a dot this is considered to be invalid and Notify
will refuse to send the email, resulting in an unactionable
`BadRequestError`.

This improves our validation to ensure these emails don't make it in to
the service in the first place. Currently in production there are 305
parents with an email address that will never send.

Jira-Issue: MAV-1961