Optimise migration that replaces `patient_session_id`
This optimises the code we use in this migration to ensure it runs much
quicker, by executing a single query. This is to avoid a situation where
the migration takes too long to run in production and causes issues with
the deployment.

This migration has already run in QA and test, but I've tested locally
and this change does the same thing. I will test this in data
replication before deploying.