Fix error when submitting invalid vaccinate form
An error page can show with a specific combination of inputs to the
vaccinate form. If the patient has multiple delivery sites, and the user
chooses that the vaccination didn't happen, and they've got other
validation errors on the form, then the user will see an error page
instead of the validation errors.

Sentry-Issue: 6869158586