Set batch team when importing offline
In c429e776f9c5485865335f6805af3dcccf2318ff we made it so that the team
on the batch when importing is set to `nil`. This makes sense when
importing historical vaccinations, but when uploading offline
vaccinations we know the exact session and team that performed the
vaccination, and therefore we should associate the batch to that team so
it appears in the teams list of batches.

Jira-Issue: MAV-1970