Simplify `Batch` look up when editing a record
This updates the `batch` method on the `DraftVaccinationRecord` model to
no longer require the batch to be in scope for the current team. This
prevents a potential issue where you're editing a vaccination record
with a batch that's not attached to any team.

This follows the approach we have already for `patient` and `location`,
and the logic is that we're not directly editing these associated
models, so it's safe to find them outside of the current team. Instead,
if a user were to edit the batch of a vaccination record, they would be
doing so by choosing from a list of known batches, so they would be
editing the `batch_id` and not any of the details on the batch itself.

Jira-Issue: MAV-1970