Ensure sync job runs after transaction
There's an issue where when a vaccination record is updated, the updates
aren't sent to the Immunisations API because the job runs before the
transaction has committed the data to the database. This commit reworks
how the callbacks work to ensure that the job will only run after the
data has been committed.

Jira-Issue: MAV-1972