Prevent Imms API records being edited by an offline spreadsheet
Vaccination records which are received from the Imms API are considered
immutable in Mavis, because we are not the vaccination record's primary
source. As such, users should not be allowed to edit the records by any
means, including an offline spreadsheet upload.

We (Thomas, Mike, Alistair) considered exposing this as a validation
error on the import, but that approach was discarded. This is because we
would have needed to duplicate a lot of the import logic to decide
whether it was appropriate to throw a validation error. This led to
circular dependencies in the validation logic.

Instead this approach was taken, where an adjusted import issue page is
shown, forcing the user to discard the incoming changes. 

Manually set `apply_changes` in the `.html`

Instead of depending on custom logic in the form.