Refactor `AppPatientSessionRecordComponent#render?`
This refactors the method to always consider whether there is a session
date today, as it should never be possible to record a vaccination if
there isn't a date today.

3f6b528c633ddee6ef043e6abc3f7e8711139f14 tried to fix this issue, but
there seems to be a scenario where the existing position of the check
on `today?` doesn't work and the the form is still trying to be
displayed (which then fails because the `session_date` is `nil`).

Sentry-Issue: 6864512518